Struct afterparty::Hub

pub struct Hub { /* fields omitted */ }

A hub is a registry of hooks

Methods

impl Hub

fn new() -> Hub

construct a new hub instance

fn handle_authenticated<H, S>(&mut self, event: &str, secret: S, hook: H) where H: Hook + 'static, S: Into<String>

adds a new web hook which will only be applied when a delivery is revcieved with a valid request signature based on the provided secret

fn handle<H>(&mut self, event: &str, hook: H) where H: Hook + 'static

add a need hook to list of hooks interested in a given event

fn hooks(&self, event: &str) -> Option<Vec<&Box<Hook>>>

get all interested hooks for a given event